restructure getScheduleUrlParam functions
